Added docker compose
Added compose.yml so that MySQL can be used with the backend.

Modified launchSettings.json to connect to the citizen-proposal-app
docker network so that it can communicate with the DB.

In the future, to start development on the backend, first use `docker
compose up -d db` to start the DB locally, then open the backend project
with VS and start the app. The backend should then be able to talk to
the DB.

Todo: Add DB users