Fixed DB connection issues and added MigrationExtensions
The suggested connection string in README was wrong. I was supposed to
use the container name as the hostname instead of localhost, since the
DB is in another container. Updated README and compose.yml to give both
containers fixed names.

Added MigrationExtensions that adds an extension method to
IApplicationBuilder that can migrate the DB on app startup.

Made it so that if the environment is Development, migrate the DB on
startup.

init.sh also didn't give the default user permission to create tables
which caused migrations to fail. It has been fixed.