Changed the workflow of starting the local test DB
EF Core migrations by default use the connection string specified by the
DbContext. In order to allow migrations, I granted extra permissions to
the default user specified in the MYSQL_USER file. However, this meant
that in production, the default user will have the permissions to alter
the schema. To only allow the default user to do CRUD operations, its
permissions have been reduced, and a user called Migrator will be
created by init.sh that has the necessary permissions to run migrations.
This means that the --connection option is now needed to use the
Migrator user when running migrations. README has been modified to
describe the new workflow.

Because of this change, the backend app no longer runs migrations on
startup, even if the environment is Development.

Added the initial migration files.