Added the instructions about starting the backend app without VS
Gave the Migrator more permissions to allow it to run migration SQL
scripts.

Added the migration SQL script from nothing to AddedPostAndTagDataSeed.

Added the development certificate used for local development purposes
and modified the csproj to make MSBuild copy it to the output directory.

Ignored the env directory since env_files can be optional and I didn't
know that.

Added the environment variables "ASPNETCORE_HTTP_PORTS" and
"ASPNETCORE_HTTPS_PORTS" to the backend container in compose.yml even
though those are the defaults for clarity.

Added the "backend" profile to compose.yml.