Added User.Loginable, Post.Author, data seeds, and a new migration
The Post.Author migration property were missing, so it has been added.
Due to this change, the migration generated actually drops the foreign
key from Post to User named UserID and adds another one named AuthorId.
This is because the navigation property was missing, so it got added as
a shadow property with the default name UserId.

Added User.Loginable so a user can be marked as unloginable. This is
added to add a "demo user" that will be the author of some posts but is
not an actual user.

Added the missing AuthorId values to the data seeds of the posts and
added a data seed of a demo user.