Added an endpoint that queries a post by its ID
Changed launchSettings.json so that the browser starts with the path
/api appended.

Changed the visibility of some classes to public because controllers
need to be public, and the classes they use also need to be public and
so on.

Suppressed a few warnings.

Made analyzers apply to all places, including non-public facing APIs,
and corrected the warnings that showed up after doing so.

Added a few DTOs.

Added AutoMapper.

Added AutoMapperProfile that configures AutoMapper.

Added AutoMapperExtensions that lets you call AssertConfigurationIsValid
in the Main method.