Changed backend app listening ports
Modified compose.yml and launchSettings.json to make the backend app
listen on ports 8080 and 8081 on the host instead. This is because it is
expected that only the frontend app or a web server will listen on ports
80 and 443 on the host.

Renamed PostController.cs to PostsController.cs.

Updated README.md to remind developers that docker compose up doesn't
automatically rebuild images.