Made the backend automatically add new tags when submitting a new post
Now, when a new post contains tags that are not in the DB, those tags
are automatically added.

Fixed a bug where when logging in, the Argon2id parameters used were
always hardcoded instead of being fetched from the DB.

Made AddPost use JSON instead of form-data because it's difficult to
send list data with forms.

In AddUser, the Argon2id parameters used are no longer hardcoded with
literals.