Added endpoints that can add and query comments
Renamed PostQueryResponsePostDto to PostQueryResponseDto,
PostQueryResponseDto to PostsQueryResponseDto, and PostQueryResponseDto
to PostsQueryResponseDto.

Clarified some documentation.

The checks for race conditions regarding user deletion were useless
since a user could be deleted after the checks were done. They were
removed, and the documentation for 401 responses was made to stop
mentioning that they could contain response bodies.

Removed the explicit returns of 500 responses because I thought it
should be reserved for unexpected situations like unhandled exceptions.
They were returned only if the session token authentication handler
didn't include a user ID in the ClaimsPrincipal, which shouldn't happen
anyway. If it can happen, then that's an unexpected situation because
that's a bug, so I also shouldn't explicitly check for it.