Added the ability to search posts with AI
Modified PostsQueryRequestDto so that it only has one Keyword property
instead of Title and Content because the AI can only search titles, and
also because I find it not very ergonomic if Title only searches titles
and Content only searches contents. If AI is not enabled, Keyword
searches for both titles and contents.

Suppressed CA1002 for private code because it was not necessary.