Fixed a bug where having an empty tag when submitting a post causes 500
Fixed a bug where having an empty or whitespace-only tag in the tag list
when submitting a post causes NullReferenceException.

Added ElementRequiredAttribute for applying RequiredAttribute to every
element in an IEnumerable.

In ElementMaxLengthAttribute, changed the element type of the foreach
from object to object? to better convey the intent, even though it
doesn't change anything.