KeywordSearch
It can now search for posts with keywords in the title or content. It can use the sorting options, tags, and author filters on the side. Currently, it seems there are no other authors or different publishing times, so I'm not sure if the author filter and publishing time sorting will work properly.